test(reorg): eval core tests => eval_spec.lua
